pyproject.toml
[project]
name = "ccpm-chatgpt5"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "matplotlib>=3.10.5",
    "pandas>=2.0",
    "openpyxl>=3.0",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.403",
    "pytest-cov>=6.2.1",
]

[tool.setuptools.packages]
find = {}
